Moving oidc_config_input to new framework
