+++ provider/oidcconfiginput/rosa_oidc_config_input_state.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package provider
+package oidcconfiginput
 
 import "github.com/hashicorp/terraform-plugin-framework/types"
