+++ provider/provider.go
@@ -36,6 +36,7 @@ import (
 	"github.com/terraform-redhat/terraform-provider-rhcs/provider/dnsdomain"
 	"github.com/terraform-redhat/terraform-provider-rhcs/provider/groupmembership"
 	"github.com/terraform-redhat/terraform-provider-rhcs/provider/oidcconfig"
+	"github.com/terraform-redhat/terraform-provider-rhcs/provider/oidcconfiginput"
 )
 
 // Provider is the implementation of the Provider.
@@ -202,6 +203,7 @@ func (p *Provider) Resources(ctx context.Context) []func() resource.Resource {
 		dnsdomain.New,
 		oidcconfig.New,
 		groupmembership.New,
+		oidcconfiginput.New,
 		// TODO uncomment this after ClusterRosaClassic resource is fixed
 		// clusterwaiter.NewClusterRosaClassicResource,
 	}
